+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,55 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.AvailableReservationTimeResponse;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.SaveReservationTimeRequest;
+import roomescape.service.ReservationTimeService;
+
+import java.net.URI;
+import java.time.LocalDate;
+import java.util.List;
+
+@RestController
+@RequestMapping("/times")
+public class ReservationTimeController {
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping
+    public List<ReservationTimeResponse> getReservationTimes() {
+        return reservationTimeService.getReservationTimes()
+                .stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> saveReservationTime(@RequestBody final SaveReservationTimeRequest request) {
+        final ReservationTime savedReservationTime = reservationTimeService.saveReservationTime(request);
+
+        return ResponseEntity.created(URI.create("/times/" + savedReservationTime.getId()))
+                .body(ReservationTimeResponse.from(savedReservationTime));
+    }
+
+    @DeleteMapping("/{reservation-time-id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable("reservation-time-id") final Long reservationTimeId) {
+        reservationTimeService.deleteReservationTime(reservationTimeId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/available-reservation-times")
+    public List<AvailableReservationTimeResponse> getAvailableReservationTimes(@RequestParam("date") final LocalDate date, @RequestParam("theme-id") final Long themeId) {
+        return reservationTimeService.getAvailableReservationTimes(date, themeId)
+                .getValues()
+                .entrySet()
+                .stream()
+                .map(entry -> AvailableReservationTimeResponse.of(entry.getKey(), entry.getValue()))
+                .toList();
+    }
+}
