+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,0 +1,65 @@
+package roomescape.service;
+
+import org.springframework.stereotype.Service;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeAvailabilities;
+import roomescape.dto.SaveReservationTimeRequest;
+import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationTimeRepository;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.NoSuchElementException;
+
+@Service
+public class ReservationTimeService {
+    private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
+
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
+        this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
+    }
+
+    public List<ReservationTime> getReservationTimes() {
+        return reservationTimeRepository.findAll();
+    }
+
+    public ReservationTime saveReservationTime(final SaveReservationTimeRequest request) {
+        validateReservationTimeDuplication(request);
+
+        return reservationTimeRepository.save(request.toReservationTime());
+    }
+
+    private void validateReservationTimeDuplication(final SaveReservationTimeRequest request) {
+        if (reservationTimeRepository.existByStartAt(request.startAt())) {
+            throw new IllegalArgumentException("이미 존재하는 예약시간이 있습니다.");
+        }
+    }
+
+    public void deleteReservationTime(final Long reservationTimeId) {
+        checkReservationTimeExist(reservationTimeId);
+        validateReservationTimeExist(reservationTimeId);
+        reservationTimeRepository.deleteById(reservationTimeId);
+    }
+
+    private void checkReservationTimeExist(final Long reservationTimeId) {
+        if (!reservationTimeRepository.existById(reservationTimeId)) {
+            throw new NoSuchElementException("해당 id의 예약 시간이 존재하지 않습니다.");
+        }
+    }
+
+    private void validateReservationTimeExist(final Long reservationTimeId) {
+        if (reservationRepository.existByTimeId(reservationTimeId)) {
+            throw new IllegalArgumentException("예약에 포함된 시간 정보는 삭제할 수 없습니다.");
+        }
+    }
+
+    public ReservationTimeAvailabilities getAvailableReservationTimes(final LocalDate date, final Long themeId) {
+        final List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
+        final List<Reservation> reservations = reservationRepository.findAllByDateAndThemeId(date, themeId);
+
+        return ReservationTimeAvailabilities.of(reservationTimes, reservations);
+    }
+}
